Dashboard/requirements.txt
dash
pyserial
plotly